Fixing GASNet communication layer for Mac OS X Lion; removing assumption
[reviewed by Sung]

Since we started using GASNet, we've been assuming in our "segment
everything" implementation that global variables will live at the same
address in every instance of the program that GASNet fires up.  We
knew that this may not be a safe assumption, but had never been bitten
by it, so never bothered to remove it.  Turns out that the failures
we've been seeing on Mac OS X Lion are a result of it -- our GASNet
launches over UDP there broke this assumption.

A workaround for this is simply to use the "fast" segment for Lion,
but it seemed worthwhile to fix this assumption while we're here, so
this commit does just that.

In segment fast, we use the beginning of locale 0's memory segment
(stored in seginfo_table[0]) to store the locations of the program's
global variables.  In this commit, we set up seginfo_table[0] in the
segment-everything implementation to point to a similar segment of
memory, which results in fewer differences between the segment fast
and segment everything implementations apart from the setup (see,
for example, the removed #if conditionals at the end of the diff).

To make this work, the setup for segment everything is a little more
complicated because we can't rely on GASNet to hand us the segment for
each process.  Instead, we have locale #0 malloc a piece of memory for
its segment (the other locales don't need a segment in this
implementation), and then broadcast this location/size (and NULL/0 for
the other locales) to all the other locales so that they have the
same seginfo[] data structure as in the segment-fast implementation.
This requires one new active message type to implement the broadcast.

The broadcast itself is done in the naive O(numLocales) manner.  The
GASNet team has some work underway to support an official
broadcast/collectives interface, so rather than doing something smart,
I'm dragging my feet in hopes that they get to it before this becomes
a performance issue for us.  I think we currently have other
O(numLocales) startup costs, so expect/hope that this won't result in
a bottleneck.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19687 3a8e244f-b0f2-452b-bcba-4c88e055c3ca