Adding some futures for writeln() calls that are passed function arguments.
I think we agreed informally over chat that passing a function to
writeln() should probably be a compile-time error.  Regardless, the
current implementation has a variety of surprising results depending
on the typ of function.

I ran into this a couple of times because I would forget to put parens
when wanting to print the value returned by a function.  I guess I am
a sloppy Chapel programmer.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19693 3a8e244f-b0f2-452b-bcba-4c88e055c3ca