Bug fix for delete from remote locale
[Code reviewed by Sung]

This change fixes a bug that Sung ran into yesterday in which calling
delete on a class instance from a locale other than the one on which
the object is allocated causes problems.  The solution used here is to
insert an "on this" clause in the normalize pass.  When compiling for
--local, we avoid doing so since it's unnecessary and would just
result in a bigger IR for the program.

I tested this fully on the normal test run and did a reasonable amount
of spot-checking key tests on GASNet, but the full test could show
other problems (though I don't expect it will).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19694 3a8e244f-b0f2-452b-bcba-4c88e055c3ca