Committing new version of Qthreads
The Qthreads team released version 1.7.1 today and it seems to be
working pretty well with as much of the examples directory as I had
time to run today (and they've done a bunch of testing as well), so
I'm committing it to get a good night's run in before the weekend
comes.

The one main change this causes for our installation is that we now
throw a --disable-spawn-cache flag at configure time in order to
disable their new spawn cache optimization which postpones adding new
tasks to the task queue until something was going to touch the task
pool anyway.  This causes problems for some Chapel codes because tasks
for begins are not inserted until the current task comes to a yield
point (as I understand it).  We talked about a strategy in which
begins were pushed onto the task queue aggressively and spawn caching
could be applied to the group of tasks created by a coforall or
cobegin, but that requires more work, so we put it off for now,
relying on this workaround instead (which causes all tasks to be
pushed proactively onto the task queue).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19695 3a8e244f-b0f2-452b-bcba-4c88e055c3ca