Retiring future that I fixed yesterday
This is the future that my commit yesterday fixed.  I didn't delete
it at the time because I wanted to see that it showed up in our
new regression testing report format in the right place (it did).

In removing it, I've also made it a comm-none test just for the sake
of completeness (Sung used the .skipif to avoid a false positive).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19700 3a8e244f-b0f2-452b-bcba-4c88e055c3ca