This commit resolves the gasnet/--no-local regressions from yesterday
With my change yesterday to insert on-clauses around delete statements,
we ran into a semantic check in the compiler that flags the use of
on-clauses within serial iterators (in the event that a delete took
place within an iterator).  Thinking about this a bit, it seems that
the case that the iterator was trying to guard against is the use of
a yield within an on-clause within a serial iterator.  

To that end, I tightened up the semantic check to check for this case
(yield within on within serial iterator) and updated an existing .good
which tested this error.  I also added two new tests to distinguish
between the error and non-error case and a .future asking whether we
should flag such errors when compiling --local (we don't currently
because all on-clauses are dropped at parse-time; and because, arguably,
the on-clause cannot be referring to any other locale for --local
compilation).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19701 3a8e244f-b0f2-452b-bcba-4c88e055c3ca