Reinforced 'locale' as a primitive type in the spec; fixed a typo.
This commit does not represent any decision re primitive types -
it just reflects more consistently what we have now.
(Which may change when hierarchical locales are added.)

* Moved the locale type from being a sibling to primitive types
to being one-of the primitive types - in the Types chapter intro,
the syntax diagram, and the Primitive Types section.

* Adjusted a spec example (Statements/typeselect.chpl)
to reflect the current list of primitive types completely.

Also replaced the plural (locale types) with singular, for adequacy.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19719 3a8e244f-b0f2-452b-bcba-4c88e055c3ca