Snapshot of Jeff Keasler & Brad's unstructured LULESH rewrite
Jeff and I pair programmed today to take the original LULESH which
used, and at times assumed, a dense 3D domain+array to store the
mesh and converted it into an unstructured case using 1D arrays
and indirect indexing.  More verification is needed before testing
is started, so this is a snapshot for now.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19723 3a8e244f-b0f2-452b-bcba-4c88e055c3ca