Added a performance test for SSCA2.
A precomp script takes the most recent version of SSCA2_kernels.chpl
from the release directory and inserts commDiagnostic calls (and
prints the results), saving the new file to SSCA2_kernels.chpl in this
directory.

The test is compiled with the RMAT graph only and run for SCALE 6
(exact betweenness centrality) using --dataParTasksPerLocale=1 to
limit variability (on 4 locales if applicable).

For CHPL_COMM!=none, the prediff script rewrites the output logfile
allowing for a +/- 1% variance in the number of gets, puts, and forks.
For CHPL_COMM=none, the script does nothing with the output.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19724 3a8e244f-b0f2-452b-bcba-4c88e055c3ca