Adding --logtmp to start_test for single-node parallelism.
start_test creates a file with the same path
(path is in the variable logtmp) even when run in parallel
on a single machine. I was experimenting with
a GNU parallel test setup, and needed to be able
to run multiple start_tests in the same directory
on the same machine. The file was causing conflicts,
so this patch just adds a command-line option
for it, so that it can be specified to something
unique when running in parallel.

Reviewed by Sung.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19731 3a8e244f-b0f2-452b-bcba-4c88e055c3ca