Optimize BlockDist bulk transfer for the case where both arrays are declared over the same domain.
Prior to this commit, for A=B where A and B are declared over the same
domain, the bulk transfer optimization resulted in a number of remote
gets (there should be 0) and 2 extra forks on locale!=0 (or whatever
locale initiated the assignment).  This commit creates a special case
in BlockArr.doiBulkTransfer() that avoids retrieving the extra data
when we know that the domains are the same and thus the data is
perfectly aligned.

BlockDist.chpl: If the domains that A and B are declared over are the
  same, invoke the bulk copy routine on the local DefaultRectangular
  arrays (locArr[i].myElems).  The call itself is wrapped in a local
  block, so the number of communications for the assignment is now 0
  (as it should be).

  Fixed up some spelling errors, changed var to const, used local (to
  a locale) variables as possible in the other cases.

DefaultRectangular.chpl: Change x.locale to x.locale.id to get rid of
  the unnecessary communication required for indexing into the locales
  array.  I found this issue when I wrapped the DefaultRectangular
  bulk transfer routine in the local block.  There was a performance
  future filed for this case earlier.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19738 3a8e244f-b0f2-452b-bcba-4c88e055c3ca