Make the default distribution type used for the distribution robustness tests BlockDist if CHPL_COMM!=none.
This change sets the default distType to be DistType.block if the comm
layer is not none.  This means that we won't be testing the default
distribution on a nightly basis, but we thought that would be fine
since default gets good coverage over the entire test suite.  This
mean that the multilocale test suite used in nightly testing will now
be testing something more meaningful.

I also had to change a couple of prediffs that wrote .good files based
on the distribution used.  They now check CHPL_COMM before assuming a
default distribution type.

FWIW, I noticed that using CyclicDist with CHPL_COMM=none fails for
reindexing/test_strided_reindex2.chpl because it gets a different
error than the one expected with CHPL_COMM=gasnet.  The test expects
an error, and since we don't run this variant nightly, I just let it
be.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19743 3a8e244f-b0f2-452b-bcba-4c88e055c3ca