Committing a couple of futures related to EOF
Michael noted on email today that he was aware that there
might be problems related with checking that we're at the
EOF for a reader channel, but didn't have a test case that
captured it.  lulesh-eof.chpl is a case that fails regularly
for me.  It tries to assert that we're at the end of a file
after we've consumed the whole thing but triggers an internal
error in QIO.  The output generated by the debugIO flag
indicates that we have consumed the whole file.

The lulesh-eof2 test is a feature request that a channel
might support such a capability by default to save the user
the (minor, but annoying) pain of writing such a routine
themselves (I know I put it off for several hours, only
to have it bite me when I wasn't at EOF as I expected to be).

Both of these tests could be retired once working by folding
the resulting capabilities into lulesh.chpl.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19744 3a8e244f-b0f2-452b-bcba-4c88e055c3ca