Adding a future for a function resolution bug.
This is a stripped down version of the tests committed by Vass in
r19625.  If a distribution is used in a non-top-level scope *and* the
distribution module declares a domain that is dmapped using that
distribution, the compiler does not properly resolve the type
constructor for the distribution (and prints no candidate notes).

The tests committed in r19625 were originally categorized as module
search path bugs.  What probably contributed to the confusion is that
when we add the new type constructor to the AST, we do not properly
update the filename/linenumber and so they are basically set to
something randomish.  Thus the error message is really confusing.  I
have a one-liner to fix this, which I'll commit soon (testing now to
make sure it doesn't mess up any line numbers elsewhere).

We should probably either update the futures committed in r19625 (and
possibly move them to a more appropriate directory) or we could just
delete them.  I'll leave that up to Vass.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19751 3a8e244f-b0f2-452b-bcba-4c88e055c3ca