Filing a bug against overloading a function with an array
This is a bug that Jeff and I ran into last week which surprised me.
Defining a function and array at the same scope doesn't cause a
conflict, and causes the array to be chosen.  In this .future, I'm
proposing it should be an error that's caught at scope resolution
time, much as defining two arrays with the same name would.  Could be
a candidate for a 5-minute fix (TM).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19754 3a8e244f-b0f2-452b-bcba-4c88e055c3ca