Add SET_LINENO() call for type constructor.
Added SET_LINENO() call so that the new call expression for the type
constructor is in the same place as the default constructor.

Found this when looking at the distribution type constructor future I
just filed.  The line numbers for the type constructor are more or
less random without the call to SET_LINENO().



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19755 3a8e244f-b0f2-452b-bcba-4c88e055c3ca