Patch-set for integrating the Boehm garbage collector for single-locale.
This almost certainly will fail horribly for multi-locale.
Besides adding a boehm-gc memory layer, we needed to fix
threads-pthreads.c to make the pthread_setspecific thread-local-storage
reachable from the thread stack; otherwise the Boehm-GC collects
it. (this is a "known long-standing bug" in Boehm-GC).

This patch assumes that the Boehm-GC is installed in the system;
ie that gc.h and -lgc are available.

Greg looked at this patch and decided it shouldn't go into
trunk because it complicates the threading-layer with support
for something that won't work multi-locale.

(based on r19758)



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/ferguson/bdwgc@19760 3a8e244f-b0f2-452b-bcba-4c88e055c3ca