Add text logging capability.
On the command line, add --log=<letters> or -d<letters> to log the AST at the
end of the corresponding pass.  Pass letters are defined in log.h.  Using -d or
--log without any pass letters specified logs the AST output for ALL passes.

For each pass, output is written to a file of the name:  <module-name>_<pass-name>.ast
Output is placed in the ./log directory by default.  
Use --log-dir=<path> to specify a different directory.
Use --log-module=<module-name> to narrow the dumping to a single module.
Use --log-symbol=<symbol-name> to narrow the dumping to a specific symbol name.
Use --log-ids to cause node IDs to be inserted into the AST dump.


Details:

The former use of log.h was obsoleted and removed from ifa_vars.cpp.

In log.h there is now a #define LOG_xxx, for each pass, where xxx is the pass
name as specified in passlist.h.  The value of these #defines is a mnemonic
character which is used to select that pass in the --log= command-line argument.
The special value NUL is specified for passes which are expected to be of little
interest to the developer (e.g., the "check???" passes).

The concept of a log level was removed.  In the previous implementation,
specification of a log letter multiple times would select a higher level of
logging, but this feature was unused.

The log flags specified by the user are private to the log module.  The value of
a particular flag is tested using the logging(int) function, whose argument is
the name of one of the #defines in log.h.

The #defines for the FIRST, LAST and RUN macros were moved inside passlist.h.

Comment indentation was adjusted in passlist.h.  The second invocation of
"prune" was renamed as "prune2", so that each pass has a unique name.  This is a
new restriction on pass names, made necessary because each non-NUL pass letter
must be unique, coupled with the fact that the RUN macro takes only one parameter.

dump_ast() and its supporting routines were derived from html_view().


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19762 3a8e244f-b0f2-452b-bcba-4c88e055c3ca