Minor modification to receive_patch as a safety precaution
receive_patch does an svn revert + update before applying the patch
which can be problematic if you accidentally apply it to a workspace
with modifications that you care about.

This commit takes a snapshot of the current diffs, saving them into
~/patches/auto_receive_patch.patch as a backup, just in case the
invocation blew away a change that you cared about (e.g., if your
CHPL_HOME variable was set to the wrong workspace).

Of course, a better solution would be to save the backups to a series
of auto patches so that you could "undo" back to a receive_patch other
than the most recent, but I was too lazy to do that and haven't
actually been bitten by this yet, so figured this minor defensive
policy would be better than nothing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19766 3a8e244f-b0f2-452b-bcba-4c88e055c3ca