in test/studies/ssca2/rmatalt: switched neighbor list representation from associative domain/array to rectangular domain/array; switched to preserving duplicate edges; implemented distributed/parallel graph construction.
This works with comm=gasnet and 'local' blocks succeed.
Edge generation is still sequential (so the edge list is always on Locale 0 I think).

Request for the compiler: eliminate the need to do the following
when I want 'local' to succeed:

      proc addEdgeOnVertex(vArg, wArg) {
        on this do {
          const v = vArg, w = wArg;  // the compiler should do this for me
          local {
            ... v ... w ...

I.e. I'd like to write simply:

      proc addEdgeOnVertex(v, w) {
        on this do {
          local {
            ... v ... w ...

Duplicate edges are enabled because we no longer have the associative
domains to tell us whether an edge is already there.
For large SCALEs supposedly the difference should be small, and
we can come back to it later.

Note: the difference in rmatalt.good shows how the histogram of
#edges -> #vertices changed when duplicate edges got allowed.
This, however, is not expected to be representative of what happens
with large SCALEs.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19770 3a8e244f-b0f2-452b-bcba-4c88e055c3ca