Committing a few changes to try and quiet down qthreads testing
In working through why we're getting so many timeouts in our qthreads
testing, Dylan Stark diagnosed the problem (related to the large-scale
oversubscription that we're doing -- multiple locales per node and
still the full number of workers) and made the following suggestions
to make it perform better:

1) Configure with --with-scheduler=nemesis

2) Set QT_AFFINITY=no in the execution environment

These changes make these changes for the purposes of the nightly
testing runs, but not for the typical user under the argument that the
typical user will not be running in this ridiculously oversubscribed
environment.

Sung was on the hook to code review this for me, but didn't get a
chance before leaving.  Since these are fairly innoccuous, and
qthreads testing is fairly obnoxious this week, I'm going to commit in
hopes of getting a quieter run before the weekend.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19780 3a8e244f-b0f2-452b-bcba-4c88e055c3ca