Remove chpl_error_noexit primitive and fix a race condition in halt.
This commit removes the chpl_error_noexit primitive from the compiler
and the corresponding calls in the runtime.  We were only using them
for the varargs versions of halt() and assert() to separately print
out the first part of the error message.  Brad hypothesized that this
was to avoid leaking more strings by concatenating everything together
into a new string and passing it to the chpl_error, but since we're
exiting anyways it doesn't seem like we should worry about it.  This
will guarantee that error messages are not interleaved.

There was also a known race condition in spinhaltIfAlreadyExiting() in
error.c.  Since we now have an atomics interface, we can fix that race
condition with an atomic test_and_set.  Note that we do not use the
atomics when compiling the launcher since our launcher is single
threaded.

[reviewed by bradc]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19782 3a8e244f-b0f2-452b-bcba-4c88e055c3ca