Optimize set-up for Block and Cyclic aliased arrays.
This change optimizes the case of aliasing where the new array is an
"exact" alias of the RHS.  For example,

        var A => B;

In this case, the domains are the same and thus the RAD optimization
cache for both arrays should contain the same data.  So rather than
recollecting the same information, simply copy it from the RHS.  This
saves us an on (remote fork) to each locale, and all the remote gets
necessary to fill in the cache.  This applies to both Block and Cyclic
arrays.

While I was in there I fixed up the array's dsiDisplayRepresentation()
to print out RAD cache information.  One test prints out the
information, so I updated the output for it including adding a gasnet
version of the .good file.

Other updates for commDiagnostics tests to follow.

[reviewed by bradc]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19784 3a8e244f-b0f2-452b-bcba-4c88e055c3ca