Cleanup:   Make --log= work as expected   Improve developer help   Auto-add trailing slash for --log-dir
I had expected --log= with no argument to mean "log every pass", but existing
argument parsing allows a space even if the '=' is present.  I modified this so
that if the '=' is present, only contiguous characters are parsed as the value
of that option.

Developer help looked weird because the section header Developer Flags was
required verbatim.  The division between user and developer flags now keys off
the prefix "Developer Flags", so one can add multiple Developer Flags sections
with subtitles.  In user mode, help output is suppressed after the first
matching header is encountered.  I had earlier split the developer flags into
Debug Options and Miscellaneous, so I turned those into subtitles.

Made my text-mode AST logging more usable by automatically adding a trailing
slash to the log-dir string.  Was getting confusing results without this, since
generated file was just getting "log" jammed on as a prefix.

Let a few formal (as in non-functional) fixes bleed in from my work on function
resolution logging.  Mainly, tightened up the interfaces on the HashFns classes
by deriving them from the base HashFns class.  This makes it more obvious that
they are implementations of the abstract base class.

Details:
    include/misc.h
    util/misc.cpp
===================================================================
Added a version of handleError which writes to a file, so I can use it to log
brief node information.  Rewrote handleError() as vhandleError() accepting a
FILE* argument.  Prints using the existing format if (file == stderr).
Otherwise, uses a one-line format.
This will be used by the resolution dumping command (--log-resolution) when it
is ready for prime-time.
 
    include/map.h
===================================================================
Made HashFns implementation classes derive from the abstract base HashFns class,
and made the derived classes match the signatures in the base class.
This change is purely formal.

    main/log.cpp
===================================================================
Want --log= to mean "log all passes".  If the option does not appear at the end
of the command line, we test for a following dash.  This is still not perfect.
I need to rework args.cpp so it treats '=' and ' ' differently.

    main/runpasses.cpp
===================================================================
Ensure that the log_dir exists for all kinds of logging.
Add a trailing slash to the log_dir name if needed.

    main/arg.cpp
===================================================================
Line 86: Change the test which selects the value of an argument.  Before, it
took the succeeding argument if it was non-NULL, otherwise it tooke the second
succeeding argument (which in that case was also hopefully NULL).  This test
actually makes little sense....  Now, we select the same argument if it has
anything left in it and the next argument if not.  If the option contained an
'=', that counts as a character, but is elided before the value is parsed.

Line 232: Changed the semantics of argument parsing, so that (for the sake of the
extraneous character check), the entire value is consumed.  (This change is
internal only; the code is simpler, but the observable behavior is unchanged.)

Improve the format of developer flag section headers, by matching against
"Developer Flags" as a section header prefix now (as opposed to the section
header in its entirety.

    main/driver.cpp
===================================================================
Improved the format of developer flags sections after removing the empty developer
flags section header.  Improved flag descriptions for the new logging feature,
to make obvious the similarities and differences with the HTML dump. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19787 3a8e244f-b0f2-452b-bcba-4c88e055c3ca