Another improvement to receive_patch
My last improvement to receive_patch tucked a copy of the current
diffs to CHPL_HOME in a backup file before reverting.  After chatting
with Sung, this commit is a slight improvement in which the auto
backup file is uniquified by the name of the patch being applied so
that if you run multiple distinct receive_patches in a short amount of
time, you won't keep blowing away the single backup patch.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19790 3a8e244f-b0f2-452b-bcba-4c88e055c3ca