Improving mechanism for requesting nemesis scheduler from Qthreads
In my original commit, I added a new Makefile variable and a
conditional in third-party/qthread/Makefile to add a new configuration
option when it was set.  On Sung's suggestion, this commmit changes
this to simply set the appropriate Makefile variable in the
cron/nightly script.  The behavior should be the same overall, it's
just a little more direct.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19797 3a8e244f-b0f2-452b-bcba-4c88e055c3ca