Added comments and refactored wrapper functions.
Pushed caching out of the functions which generate the wrappers and into the
functions which figure out whether wrappers are required or not.

In default_wrap(), reorganized the main IF in the loop which walks the formal
argument list -- so it is clearer that the default behavior is to use the
wrapper formal verbatim.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19803 3a8e244f-b0f2-452b-bcba-4c88e055c3ca