Refactored program initialization so that more is done within the first task
[reviewed by Greg Titus]

This commit refactors how programs are kicked off so that more work is
done by the first Chapel-created task rather than the C thread that is
executing main().  In particular, the first calls to
chpl_task_setSerial() and to initialize the standard modules are now
done within a task rather than within main().  The original motivation
for this was to reduce the need for a tasking layer to support
chpl_task* calls (like chpl_task_setSerial() above) outside of a task
that it created.  It also results in a slightly cleaner code
organization, arguably.

As part of this change, I added a new optional callback macro to the
tasking layer called CHPL_TASK_STD_MODULES_INITIALIZED() which will
default to a no-op if not overridden by a tasking layer.  This permits
the tasking layer to execute some code after Chapel's standard modules
have been initialized -- for example, we use it in the fifo tasking
layer to register the main task in the task table (since, as of this
change, the task table will not have been created once the main task
creates itself).

Most of the code changes related to this change are in renaming.  I
re-used the name chpl_main() for the routine used by
chpl_task_callMain() which now lives in runtime/src/main.c (whereas
before it was generated by the compiler).  I called the
compiler-generated main routine chpl_gen_main() to make it clearer
that it was generated code, and also renamed the compiler variable
that refers to this function from chpl_main to chpl_gen_main
(accounting for most of the compiler diffs).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19818 3a8e244f-b0f2-452b-bcba-4c88e055c3ca