Filing a bug against incorrectly propagating initializers amongst configs
This test shows that if a config var is overridden on the command line,
that value is incorrectly propgataed to other configs with which it
shared its initialization expression.  This is due to a bad code
transformation early in compilation that initializes each config in
terms of the previous variable.  I've got a simple fix that I'm going
to test, though I'm nervous that it's not correct.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19828 3a8e244f-b0f2-452b-bcba-4c88e055c3ca