Incorporating Qthread improved native sync vars into Chapel tree
In the Qthreads source tree, r3046 added a capability for more direct
sync var utilization that was originally designed to make use of
native MTA full/empty variables.  This commit merges that change into
our 1.7.1 snapshot of qthreads and enables it by adding qthreads to
the conditional in ChapelBase.chpl that looks for this fix.

Note that this is still not the optimal solution -- it would still be
preferable to (a) narrow sync vars to only the basic scalar types as
intended and then (b) narrow the implementation to not rely on the
generic class approach to implementing sync and single, along with (c)
beefing up support for initializing external types to the extent that
we could refer to qthread's full/empty variables directly.  That's
work that we haven't been able to prioritize yet, though...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19833 3a8e244f-b0f2-452b-bcba-4c88e055c3ca