Updating another incorrectly passing future
This is another future that started passing as a result of the
int->64-bit change.  This one also assumed default 32-bit ints, and I
also incorrectly updated the .good file as part of my massive
query-replace.

While I'm brainstorming about what to do about these futures, this one
seems like it could be handled due to the fact that the values are
params.  All param math could be done on maximally-sized integers and
then safely downcast back to the original type if appropriate (which
would result in a compile-time error here).  

Or, as I mused about in my log message for the int->64-bit change
(note 1a), we could always do param math on the largest size
ints/uints which has the upside of not losing precision and being
automatically coercible back to smaller integer sizes, but the
downside of always promoting.  In that case, the .good here would be
correct.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19837 3a8e244f-b0f2-452b-bcba-4c88e055c3ca