Fix performance bug with ref counting using atomics.
For the list lock in the distribution and domain classes, if the
acquire happened on a different locale than the one that the class was
allocated on, each test of the lock resulted in an on to the remote
locale.  Highly undesirable if there is contention.

Since we know that we will need to do at least one remote fork for the
lock, we might as well wrap the remove_* and add_* code in an on
clause.  The lock function is only called within the remove_* and
add_* so rather than adding a redundant on, I just put a comment in
there that warns that if the lock is to be acquired from a remote
locale, we should wrap it in an on clause.

Also, updated a test that expected an error that included line numbers
from ChapelDistribution.chpl.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19854 3a8e244f-b0f2-452b-bcba-4c88e055c3ca