Added feature request futures on allowing assignment to 'this' and on propagating constness from a variable of a class or record type to its fields (and further).
These are a follow-up to a discussion between Brad and Vass, but the
motivation and some ideas have been discussed elsewhere, too.


Example from constnessOfThisInMethods.future:

    record R { var x: int; }
    proc R.mod()   { x = 5; }
    proc R.nomod() { writeln(x); }

    var r1: R;
    r1.nomod();
    r1.mod();

    const r2: R;
    r2.nomod();
    r2.mod();  // compiler must report error here; the previous calls are fine
    r2 = r1;   // correctly disallowed today

Rationale: in my mind, 'const r2' means the program is disallowed
to modify r2's value. The latter means exactly modifications of r2's fields.


From propagateConstnessOfVariableToFields.future:

Today the const-ness of a field is the same for all instances of the class.
In some cases it is limiting: instead, we'd like a field to be constant
if the instance is referenced by a 'const' variable, otherwise we'd like
the field to be variable.

Const-ness of a record field declared as 'var' already follows the constness
of the variable that stores the record.
However, even within a record, it may be desirable to propagate its
constness into its fields. E.g. if the record has the field myC that
is of a class type C, we may want the constness of C's fields to
match the constness of the myC field that references the C instance.


Both .futures also sketch some ways to provide these features.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19855 3a8e244f-b0f2-452b-bcba-4c88e055c3ca