Fixing class ID regression
[reviewed by Sung]

When changing default 'int' to 64-bits, a subtle regression creeped
in, in which a 64-bit temp was used to store the dynamic class ID of
a remote class when remote dynamic dispatch was executed.  Under
certain conditions, this caused problems.  After discussing options
with Sung, we decided to change the enumeration used to store dynamic
IDs in classes into a fixed-width integer in order to remove uncertainty
in the C code about how wide the type is.  This commit changes it to
a 32-bit int.

We talked about having the compiler automatically use a class ID that
was the minimum required width since this ID takes up space in every
object, and in most cases, an 8- or 16-bit int would be plenty.  However,
in the end we had concerns about what the alignment of this value would
be and whether that might cause communication problems for one-sided
communication.  This may have been overly conservative, but given the
close timing to the release date, it seemed prudent for the time being.

The main downside to this change is that we lose some of the symbolic
elegance in debugging of not having the enum value.  If this becomes
an issue, we discussed introducing a routine that would map class IDs
back to strings.  Please feel free to request this if/when it would be
of use to you.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19860 3a8e244f-b0f2-452b-bcba-4c88e055c3ca