Updating REGRESSIONS based on past few days' work
I stupidly svn reverted some changes one night in a bad moment, so
this is missing a day or two of entries that I haven't worked up the
stomach to re-enter.  Since I owe a review of this file anyway for the
release, I'm dragging my feet on that for now.

In better news, though...

+ types/range regressions due to int->64 change resolved

+ linux32 int->64 regressions fixed

+ qthreads working much better with new config options and syncvar
  implementation

- new nondeterministic failure in the CCE compiler when compiling
  spec/Generics/query.chpl

* more valgrind issues due to unhandled syscall (waiting on valgrind upgrade)

* normal nondeterminisms



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19866 3a8e244f-b0f2-452b-bcba-4c88e055c3ca